requirements.txt
requests>=2.31.0
psutil>=5.9.5
dnspython>=2.4.0
scapy>=2.5.0
rich>=13.5.0
